pyproject.toml
[tool.poetry]
name = "summarizer"
version = "0.1.0"
description = ""
authors = ["Dolf <dnoordman@schubergphilis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dspy-ai = "^2.6.2"
dspy = "^2.6.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
